requirements.txt
langchain-groq>=0.1.0
langchain-core>=0.1.0
fpdf>=1.7.2
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.25.0
scikit-learn>=1.3.0